fix error when handling a file that is only whitespace and C pre-processor stuff
In kind of a karma move, this might only happen with strange files like
armyknife-lib-impl.c which is essentially ALL C preprocessor directives,
comments, and white-space.

The karma part comes in because I'm the one who decided an initial
capacity for buffers (and other structures) was a reasonable idea though
of course I'm now seeing this as a bad interface and I guess I will just
change it!