parse storage class specifiers for global variables
This actually seems to allow us to now fully parse out own source code
(modulo function bodies and a few forms where we cheated). This means we
should be getting ready to be able to generate header files and then we
will have a better sense of how far along we really are.