finish removing prototypes
I can see this will be a somewhat painful process at this point. Really
what we probably want is to have two development trees where one is a
little older and generally works. The right calls to make
save-bootstrap-prototypes could mitigate things a bit. The best thing is
to give this a try and adapt.