move extractions functions from main.c to header-file-extractor.c
Probably not the best name but it might itself get turned into multiple
files later. I'm still fooling around with using the Nth version the
develop the Nth version rather than using the N-1th version checked out
in a different directory or something like that. I'm only doing this
because I feel like I am learning something in the process but I'll
probably switch the having a "stable" version soon as that will require
less mental overhead to get things just right.

One interesting thing about this checkin is that header-file-extractor.c
doesn't depend on c-single-source-file/generate-header-file. That's
really only because it is only composed of top-level functions. In case
it isn't obvious, c-single-source-file/generate-header-file was an
intermediate tool. It will be interesting to see if anyone if using it
but by that time, omni-c will do that same thing but better.