start wiring in extract enums
This time to get the prototype to work I correctly, I put it in the
non-top down order, then moved it to where I wanted it after
generatating the prototype header files. I still think just having a
known good tree and a development tree is the right way to go.