start designing a symbol table
This will help us implement the ordering algorithm for structures.
Symbol tables of some sort (also called environments in some languages)
are kind of inevitable so I don't mind starting to work this into the
project at this time even though we don't technically need it for things
besides typedefs and structures yet.