fix additional problems related to token_to_string signature change
I certainly could have found this before checkin.

  make save-bootstrap-prototypes
  make clean
  make bootstrap
  make

I should have caught it though maybe I would have to have hand editted a
generated file again. Since this are human readable and editable, if
someone developing omni-c occasionally resorts to this, it's not the end
of the world since we are using version control.