start messing around with token based manipulation
I originally thought of this to allow adding Java style annotations to
the source code without needing to worry about them in the parser by
converting the token to a whitespace token (which will then get filtered
out). Could be an interesting thing to allow people to hook into the
compilation process to do things the CPP can't easily do.