we can now output "metadata" for enums and have it compile
So 60 lines of code or so and we just added some runtime reflection
capablities to the C language!

Structure reflection shouldn't be that much harder and I believe that
will accelerate development as I won't have to write "debug printers"
anymore.

As always, figuring out how to use this in the boot-strapping process
means I won't rely on it right away (or so I say, laziness is the mother
of abstraction).