update run-tests.lua to work on C files
We can still add in a .sh file if we need it be automatically invoke the
correct bash script based on the contents of the C file. Also rename
test files to just .c - I don't think the .oc suffix will ever catch on
for me and hence likely not for others either.