add sub-expression parsing and fix bug in precedence checking
Obviously we need golden files for our tests. Some of them are still
wrong like postfix operators (I think we are just missing the
instruction handling for them). We need golden files for them as well as
we can probably "expect" a semi colon after them to make sure we parse
the entire "file" when doing expression tests.