add support for sizeof
we can add cast(a, b) more of less the same way but I'm not sure how to
add C style casts. (int)a is not the same as a sub-expression followed
by an identifier. (*foo)(a, b, c) is a call not a cast...

I think the next thing to handle is calls.