introduce a printer abstraction to c-file-printer
The basic need is to carry the indentation level when doing recursive
printing of elements inside a function which we are ready to start doing
now. The good news is that this accomplishes this and produce slightly
less garbage along the way as we no longer turn each token to a string.
The bad news is that sometimes the code is actually longer than using
buffer_printf.