the omni-c compiler thinks it can handle it's own source code now
While unit tests pass and a self.c is generated, it doesn't compile.

There are still some shortcuts I known about like
parse_balanced_construct but it's not clear if any of the current
shortcuts are the actual problem right now since we aren't trying to
fundamentally change the code within function bodies so something is up
with the parser or the printer (less likely to be the lexer but who
knows).

Several strategies going forward are possible:

1) debug self.c compilation issues and trace back to a parser or printer
issue or whatever (adding more tests)
2) focus again on c-armyknife-lib compilation since maybe something pops
up there that fixes self compilation of omni-c - a bit of a long shot
but it's still forward progress in a way
3) see if we can transpile a single file at a time and link it with the
rest of omni-c which would have positive long term benefits for
debugging going forward
4) add additional unit tests perhaps by mining test cases from various
sources like gcc, clang, etc. even if we have to modify them a bit
5) attempt to compile other code bases - we'll likely stumble upon bugs
from another vantage point where they may be easier to fix though we are
still different enough from C that this won't be trivial in general
though it will greatly increase coverage in a certain way
6) work on random code generation in hopes that fixing a few issues that
appear simple actually unlock what could be the issue with self
compilation
7) just examine the output code and look for anything that looks funny
8) search for TODO and FIXME in the code as already written
9) do some much needed cleanup hoping something pops up
10) etc.

#3 is kind of the right path to proceeed with if one has a rock solid
interpreter and the interpreter could call compiled code making it easy
to mix and match.

Before we compiled at the expression level, self.c worked fine when we
verbatim copied function bodies. We could introduce a dial based on
hashing or a whitelist and slowly ramp up how much of our own source
code we parse at the expression level.

Most likely I will try #1 for a while, taking breaks to do #2, and have
some version of #3 in mind. #7 and #8 should be done anyways.

#9 which I very much want to do, feels like it will be easier to do
after we self-host. In terms of parsing using a hand written parser,
I've already encountered some akwardness like types wanting to parse
literals but literals wanting to parse types - perhaps the major goal of
omni-c has been to make this issue non exisisten withing a compilation
unit which omni-c allows you to define rather than 1970s era minimalism.