fix path to self.c in Makefile
Of course gcc couldn't compile self.c since we moved it. As worried
about in the previous checkin, we have work to do but now we can just
fix the issues we are seeing and cross our fingers that at least for now
we don't need the more radical measures (some of which we need to think
about longer term anyways which is why I wrote this up).