fix the pratt parser
We now leave room in the precedence so we can add one without
overflowing into the next precedence level. Also we change from
subtracting the precedence when right to left to add to the precedence
when left to right which seems to pass the newest test cases and the new
failures turn out to be bad golden files which are now fixed.