add use-statement-parser=true to the selfself target
Amazingly, this means that we are probably not too far from being able
to self-host from a checked in self.c rather than bootstrap-saved, etc.

We still need to deal with
  armyknife-lib-impl.c (which means dealing with c-armyknife-lib)
  lexer-to-string-helper.c
  to-string-helpers.c
  reflection-header.c

We also want to make sure that self performs the same as omni-c on the
200 hundred or so tests we have now. Even then it may be premature to
move to that style of self-hosting because the reason to do so would be
to allow lots of cleanup. etc., and it would be very difficult to undo
such a change.