add PRECEDENCE_POSTFIX to fix post-increment
This along with some work-arounds allows omni-c to compile itself to
self.c after fully parsing function bodies (we had success with self
targets using top-level functions only and not handling c-armyknife-lib
source files). This version of omni-c is able to pass the expression and
statement tests and is even able to compile itself again (needs more
testing). It can't yet parse it's own output (which is legal C code)
probably because we dropped the "hints" like cast() and block_expr().
Anyways, on track for a milestone and eventually being able to use a
preserved self.c to bootstrap development.