change two instances of log_warn to log_info in pratt-parser.c
While these logging statements were helpful on very small inputs, they
definitely are not that helpful anymore. I do most of my development on
a "slow" N100 based mini-pc within Emacs and getting rid of excess
logging either actually makes it faster or somehow makes it feel faster.
I'm too lazy to measure it. Certainly it's worth it to get rid of stuff
that wouldn't be useful to an end-user.

I'm not sure what the official terminology is for the parser that I have
written is.

Hand written comes to mind first. I'm not saying it couldn't be
generated from a grammar (perhaps a PEG like grammar) but it's
complicated and I'm probably going to want some conditional parsing
based on a flag at some point so PEG isn't a perfect choice right now.

Except for pratt-parser.c, it's mostly recursive decent.

Again this parser is PEG like in that certain ambiguities (I'm sure
there are some -- perhaps many) are resolved by the order in which we
attempt to parse things. (I don't think I was too clever but randomly
swapping stuff might affect things.)

It's decidedly NOT PEG right now because we don't memoize things like an
efficient PEG compiler might do (which might be wasted effort, a very
long argument if I have to present facts and data).

My goal is say parse 10K LOC/s on a single core of an Intel N100 CPU but
also utilize at least one CPU core per file for processing. I am
probably not there yet but I've only focused on correctness and
expressibility. What I have now is not the ugliest stuff even though it
is written in C! A correctly functioning program can be optimized but
it's kind of pointless to optimize a non-functioning program.

So my parser clearly back-tracks - which is going to be crucial to
proper error reporting. (The simplest rule is to present the longest
match will probably be what we focus on but I may present multiple error
paths because trying something new is fun especially visualizations.) I
may even try machine learning techniques to improve error messages if I
have to. I've already taken the error mechanism from my previous non-C
compiler, which was based off of Elm (widely regarded as having great
error messages), so I'm ready to plug better stuff into that.

A very clever sequence of tokens will probably make my parser extremely
non-linear (aka slow - kind of like how I made Netscape Navigator really
barf of a fairly small "page" back in 1996 or so because it had a very
loose policy in how it parsed comments - that's was fun in a way, but
not really realistic).

The pratt-parser is more table driven than most of the rest of the code
is (as evidence at how easy it was to correct the flaw in postfix ++).
It's a big chunk of code. I sense I can rewrite it to be much smaller
and perhaps more efficient (omni-c will eventually allow switch on
c-strings or buffers which will probably make it faster and source code
more compact at the same time?)