add a dev centric benchmark
I learned way too much doing this about the abundance of shells and BS
like that. Maybe the correct solution is to use {} with time in Make. I
don't know.

Because our compilation of noisy as hell, running from within emacs
shell takes 18 seconds on this N100 PC while running from within
gnome-terminal on the same N100 PC is 3.9 seconds so Emacs is probably
adding 12s of overhead. Ouch.