update omni-c-stable.c
Updating this file from now on should always occur in it's own commit.

This is mostly picking up a few changes from log_warn to log_debug.
There's still too much left. If I had to guess off the top-of-my-head,
the lua test running is throwing away stderr which is why this hasn't
seemed that annoying (of course the boot-strapping process is really
different, definitely my next guess - it's possible I didn't account for
that.)