move build/omni-c.c to omni-c-stable.c and update the Makefile slightly
From this point on, we should no longer depend on c-armyknife-lib.h
being available as a system include (or at all), and we just require a
single stable C file to build the bootstrap compiler. We also get rid of
the reflection-header-stable.c file and it's now should be much easier
to modify the reflection header more easily though we may still need to
change the Makefile to not require an explicit step to build generate
reflection-header.c.