update TODO
Most of the new stuff isn't significant but that's OK.

I don't think writing linearize() right now is going to be that useful
but it lays the ground-work for most other real compiler improvements
like overloading (which is probably easier to support than templates,
our other near term feature). It might not be that hard to generate wasm
if we simply write linearize(). If we then have an internal wasm
interpreter then in theory we can do stuff like comptime and macros
safer than say compiling to C and loading up dynamic libraries.