update omni-c-stable to use the Hans-Boehm garbage collector
It's not clear how compatible Hans-Boehm is with non-x86 architectures
though presumably ARM isn't a problem given how popular it is. I'm not
going to refactor the Makefile to build with both allocate.c and
gc-allocate.c right now. Since we haven't been especially cautious with
freeing things in omni-c or it's library, this is a very positive
change.

Setting the environment variable GC_PRINT_STATS=1 shows the gc kicking
in from time to time and claiming it's freeing memory.