get tests working again after making output files read-only
Add an additional header to generated output files.

Even with "make clean" some tests that write into /tmp/ were still
failing. This caused me to rethink things a bit since if making output
files read-only was such a good idea, other build systems would surely
do it. I'd hate to accidentally over-write a an important file when used
from the command line.

io.c get's another new function to make a file readable.

This may not garner me a lot of sympathy and call into question things,
but build/omni-c.c doesn't seem as "up to date" as "self.c" and I'm not
sure why yet though it's felt like this before.

I'm breaking my own rule to not update omni-c-stable.c in the same
commit because omni-c-stable.c was not able to run it's own tests twice
in succession and this is the fix for that.