add preliminary support for C style structural literals
This still needs unit tests but kind of as expected, this version
doesn't break anything.

(Perhaps it was a mistake to not add this much earlier. I was less
concerned with being "compatible" with C when I started than I am now
(since otherwise there is no reason to not just use zig which has become
much more popular since I started.)