add line directives to essentially all statements
Adding these to functions went pretty well a while ago, so this is
simply the next logical step.

We may want to add line directives for variables, structure/union and
even macro definitions but I don't understand if this is important or
not right now.

As for expressions - we've probably definitely narrowed those down to a
few lines since sane folks don't write expressions that span too many
lines (though for statements seems like a place to consider first). I
was thinking about waiting on expressions until we do "linearization"
though that is going to add quite a bit of size to the ouput C file (-
our bigger issue is that our output files are larger than I would like
for hello word. I'm not sure I need to fix that for an MVP).