update omni-c-stable.c
We clearly have to do this twice so this is the first time.

Even though not much code has changed, the next one will have lot's of
diffs because of the vast number of added line directives for each
statement though then subsequently, it won't be quite so bad.. if we
only work on a few files.