Add the basic outline of the linearizer
We can't start to guestimate that it might end up being between 500 and
1000 lines of code.

We are probably commiting to only having a single loop form (essentially
an infinite do/while(1) loop. Howewver, we don't lose "block structure"
by doing so which could be important for compilers with certain kinds of
register allocation algorithms. Technically labels/goto already crack
the fiction of C block structure however we don't have to exactly
shatter it and "99% of C code don't use use labels/goto" (source, I made
it the F up!) so even the gcc and clang optimizers are probably not
going to be as tuned to this type of input coming from the frontend.
Therefore, though part of me is saying "just go to bblocks", I'm not
sure I want to do it in this phase.

We probabably want a flag that will introduce the linearizer slowly but
we still need parse tree golden "unit tests".