start the command in it's own file pattern
My large main.c was partially the result of not being to refactor so
easily but now multiple LLMs wanted something like this and I agree it's
better.

Purposefully put archive-command.c after main.c in the Makefile to
emphasize that "order doesn't matter" (order does matter, to humans,
top-down style is literally the best option though trillions of time
I've put a function above it's caller because I've too lazy to write a
prototype -- and the order of inline functions still matters (it's
complicated - soon I will fix that).