add obvious buffer_read_file() function
I was trying to keep the total number of library functions down but
helpers like there are kind of nice and we aren't trying to be a single
file header file anymore. To get binary size down to acceptable levels,
we do have to eliminate unused functions when building at some point.