start using the built in "test" command for expressions parse/print tests
Get rid of the unused ".e" tests that were written before we had C
output. The longer format is hopefully less likely to be misinterpreted.

(Because these tests are noisy, it's easy to see things are working
correctly.)

We could probably try to move diff'ing and stuff inside of omni-c and
chip away a bit at tools/run-tests.lua a bit until it's a very minimal
test runner written in omni-c.