finish the first pass conversion over tests/lib to use omni-c test
The next steps are:

1) get rid of .sh files that are simply invoking "small tests"

2) figure out how to deal with the remaining .sh files - in theory we
can leave some of them there for a while but if there are small tests
hiding in there, maybe split the tests into large and small.

3) maybe try to compile and run all of the small tests together? Without
a solid recovery strategy (or a fancier strategy like test skipping),
it's kind of hard to "continue".

4) we can still improve how successful and failing tests present
themselves. this is certainly more important as other folks attempt to
write unit tests, etc.