update omni-c-stable.c to the latest version that can run tests
As is normally the case, lots of spurious diffs around line number
changes. Considering that most self-hosting compiler would perhaps be
using checked in binaries at this point, we are still doing OK.

Still, I'm thinking of several possible strategies (for example,
omni-c-stable.c doesn't necessarily need any source file line
information so we would strip it out with a script or a command line
argument when generating it - we could also make a convenient diff util
just to make it easier to see what the real diff is but keep the info in
the git version, etc.) Kicking the can own the road doesn't seem like a
terrible option right now to get to our magicaly 0.1 release where we
can start sharing on social media to attempt to get buy in.