small update to the linearizer which still isn't close to being ready to test
Our source line number information probably made this checkin harder
than it needed to be though the bad design of node-list.c along with an
initial incorrect usage was the "irritant" (and gcc's error message was
a bit cryptic). I updated the TODO to fix part of this.

parse-node->lowest-line-number (> 0) might kind of "fix" the debug
information for now and we'll need it anyways once we linearize (though
once linearized, that function would be cheaper...)