kludge in redundant line numbers for the for loop parts
We introduced newlines to make the output code of for loops more
readable / canonical but this tends to make stuff off by one or two.
THIS shouldn't actually matter much but it looks like we also don't put
out line directives for variable declartions which are quite common so
we'll fix that. Luckily *I'm* dog fooding this as in some ways source
line numbers actually made stuff worse for fixing issues.

(I wonder if on error, we automatically regenerate without ANY line
number information so we can see the "real" error the C compiler is
seeing..)