Start support for macOS (with dependecies provided by brew)
macOS 15.4.1 on M4 silicon, we seem to build a potentially functionaly
"omni-c" but fail during testing (not enoough output on failure to
understand this failure yet).

The deltas so far have been with "make" and "paths" to homebrew stuff
and some GNU and BSD basic tools not being compatible. Luckily the world
is being rewritten in Rust:

https://xkcd.com/927/